chore: rename attendeeRepo to event-attendee-repo
