+++ TODO.md
@@ -1,5 +1,16 @@
 # TODO
 
+## Refacto
+
+- [ ] Move to Node 24
+- [ ] Move to GitHub action
+- [ ] Check if possible to run int-test in parallel
+- [ ] Repository export single const tokens for  injection
+- [ ] Move repositories to database folder
+- [ ] Test int don’t go with drizzle neither database service —> directly queries from native pg client
+- [ ] Move Attendee to Event folder
+- [ ] Move Item to Wishlist folder
+
 ## General
 
 - [x] Backend: refacto mailer templating. Try react-mail or MJML
