+++ apps/api/src/attendee/application/command/index.ts
@@ -1,2 +1,3 @@
 export { AddAttendeeUseCase } from './add-attendee.use-case'
 export { DeleteAttendeeUseCase } from './delete-attendee.use-case'
+export { OldDeleteAttendeeUseCase } from './old-delete-attendee.use-case'
