+++ apps/api/src/attendee/application/index.ts
@@ -1,4 +1,4 @@
-import { AddAttendeeUseCase, DeleteAttendeeUseCase } from './command'
+import { AddAttendeeUseCase, DeleteAttendeeUseCase, OldDeleteAttendeeUseCase } from './command'
 import { AttendeeAddedUseCase } from './event'
 
-export const handlers = [AddAttendeeUseCase, DeleteAttendeeUseCase, AttendeeAddedUseCase]
+export const handlers = [AddAttendeeUseCase, DeleteAttendeeUseCase, OldDeleteAttendeeUseCase, AttendeeAddedUseCase]
