+++ apps/api/src/attendee/domain/command/delete-attendee.command.ts
@@ -1,14 +1,16 @@
-import type { AttendeeId, ICurrentUser } from '@wishlist/common'
+import type { AttendeeId, EventId, ICurrentUser } from '@wishlist/common'
 
 import { Command } from '@nestjs-architects/typed-cqrs'
 
 export class DeleteAttendeeCommand extends Command<void> {
   public readonly currentUser: ICurrentUser
   public readonly attendeeId: AttendeeId
+  public readonly eventId: EventId
 
-  constructor(props: { currentUser: ICurrentUser; attendeeId: AttendeeId }) {
+  constructor(props: { currentUser: ICurrentUser; attendeeId: AttendeeId; eventId: EventId }) {
     super()
     this.currentUser = props.currentUser
     this.attendeeId = props.attendeeId
+    this.eventId = props.eventId
   }
 }
