+++ apps/api/src/attendee/domain/command/index.ts
@@ -1,2 +1,3 @@
 export * from './add-attendee.command'
 export * from './delete-attendee.command'
+export * from './old-delete-attendee.command'
