+++ apps/api/src/attendee/domain/index.ts
@@ -1,4 +1,4 @@
-export * from './attendee.model'
-export * from './attendee.repository'
 export * from './command'
 export * from './event'
+export * from './event-attendee.model'
+export * from './event-attendee.repository'
