+++ apps/api/src/attendee/infrastructure/event-attendee.mapper.ts
@@ -1,14 +1,14 @@
 import type { AttendeeDto } from '@wishlist/common'
 
-import type { Attendee } from '../domain'
+import type { EventAttendee } from '../domain'
 
 import { userMapper } from '@wishlist/api/user'
 
-function toAttendeeDto(model: Attendee): AttendeeDto {
+function toAttendeeDto(model: EventAttendee): AttendeeDto {
   return {
     id: model.id,
     user: model.user ? userMapper.toMiniUserDto(model.user) : undefined,
-    pending_email: model.pendingEmail || undefined,
+    pending_email: model.pendingEmail,
     role: model.role,
   }
 }
