+++ apps/api/src/event/domain/repository/index.ts
@@ -1,0 +1,1 @@
+export * from './event.repository'
