+++ apps/api/src/event/infrastructure/index.ts
@@ -1,2 +1,3 @@
 export * from './event.mailer'
 export { eventMapper } from './event.mapper'
+export * from './event.module'
