+++ apps/api/src/repositories/repositories.tokens.ts
@@ -1,5 +1,5 @@
-export const ATTENDEE_REPOSITORY = Symbol('ATTENDEE_REPOSITORY')
 export const EVENT_REPOSITORY = Symbol('EVENT_REPOSITORY')
+export const EVENT_ATTENDEE_REPOSITORY = Symbol('EVENT_ATTENDEE_REPOSITORY')
 export const SECRET_SANTA_REPOSITORY = Symbol('SECRET_SANTA_REPOSITORY')
 export const SECRET_SANTA_USER_REPOSITORY = Symbol('SECRET_SANTA_USER_REPOSITORY')
 export const USER_REPOSITORY = Symbol('USER_REPOSITORY')
