+++ apps/api/src/secret-santa/index.ts
@@ -1,3 +1,2 @@
-export * from './application'
 export * from './domain'
 export * from './infrastructure'
