+++ libs/api-client/src/services/index.ts
@@ -1,10 +1,10 @@
-export { AuthService } from './auth.service'
-export { UserService } from './user.service'
+export { AdminEventService } from './admin-event.service'
 export { AdminUserService } from './admin-user.service'
 export { AdminWishlistService } from './admin-wishlist.service'
-export { WishlistService } from './wishlist.service'
-export { AdminEventService } from './admin-event.service'
+export { AuthService } from './auth.service'
+export { EventAttendeeService } from './event-attendee.service'
 export { EventService } from './event.service'
 export { ItemService } from './item.service'
-export { AttendeeService } from './attendee.service'
 export { SecretSantaService } from './secret-santa.service'
+export { UserService } from './user.service'
+export { WishlistService } from './wishlist.service'
